refactor out the find_files function
